% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{Combin}
\alias{Combin}
\title{Generate all possible combinations}
\usage{
Combin(exp_mean_cent, size)
}
\value{
a matrix of possible combinations.
}
\description{
Technically this returns the maxtrix of possible combinations provided the number of items and the combination size.
The number of items and the combination size are inputted.
}
\examples{
Combin()
}
\seealso{

}
