% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{ExpMeanCent}
\alias{ExpMeanCent}
\title{Calculate a matrix of MaxDiff utilities which are mean-centred and exponentiated}
\usage{
ExpMeanCent(df)
}
\value{
a matrix of MaxDiff utilities which are mean-centred and exponentiated.
}
\description{
Technically this returns the maxtrix of MaxDiff utilities which are mean-centred and exponentiated.
MaxDiff utilities at the respondent level are inputted. There are 2 calculation steps:
computing mean-centred utilities and after that exponentiating these scores.
}
\examples{
ExpMeanCent()
}
\seealso{

}
