% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{PreShare}
\alias{PreShare}
\title{Calculate preference shares of items at the respondent level and at the aggregate level, and rank them at the aggregate level}
\usage{
PreShare(df, choice_item, weight = NULL)
}
\value{
a matrix of preference shares at the respondent level as well as the aggregate level.
}
\description{
Technically this returns the matrix of preference shares of items at the individual level,
their preference shares and ranking at the aggregate level. Their preference shares and ranking
at the aggregate level are saved in the 2nd last row and the last row, respectively.
MaxDiff utilities at the respondent level, the number of items in a choice card are inputted. 
If weighting is required, the weighting matrix is also included.
}
\examples{
PreShare()
}
\seealso{

}
