% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{Rank}
\alias{Rank}
\title{Support to locate the combination with the n-highest reaching probability}
\usage{
Rank(n)
}
\value{
the combination number with the n-highest reaching probability
}
\description{
Technically this returns the combination with the n-highest reaching probability.
The maxtrix of exponentiated mean-centred utilities for possible combinations and the rank you need to extract are inputted.
}
\examples{
Rank()
}
\seealso{

}
