% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{RankPos}
\alias{RankPos}
\title{Search for the position of the combination with the n-highest reaching probability}
\usage{
RankPos(df, n)
}
\value{
the combination number with the n-highest reaching probability
}
\description{
Technically this returns the combination with the n-highest reaching probability.
The maxtrix of exponentiated mean-centred utilities for possible combinations and the rank you need to extract are inputted.
}
\examples{
RankPos()
}
\seealso{

}
