% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{ReachProb}
\alias{ReachProb}
\title{Calculate reaching probabilities of possible combinations}
\usage{
ReachProb(exp_mean_cent, combi, choice_item, weight = NULL)
}
\value{
a matrix of reaching probabilities of possible combinations.
}
\description{
Technically this returns the matrix of reaching probabilities of possible combinations,
their reaching probabilities and ranking at the aggregate level. Their reaching
probabilities and ranking are saved in the second last row and the last row, respectively. 
The maxtrix of the exponential mean-centred utilities and the number of items in a choice card are inputted.
Also, if weighting is required, the weighting matrix is included.
}
\examples{
ReachProb()
}
\seealso{

}
