% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{TopPos}
\alias{TopPos}
\title{Locate the 5 winning combinations with the 5 highest reaching probabilities}
\usage{
TopPos(df)
}
\value{
the positions of the 5 combinations with the 5 highest reaching probabilities
}
\description{
Technically this returns the 5 winning combinations with the corresponding 5 highest reaching probabilities.
The maxtrix of exponentiated mean-centred utilities for possible combinations is inputted.
}
\examples{
TopPos()
}
\seealso{

}
