% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdturf_function.R
\name{TopUti}
\alias{TopUti}
\title{Extract the reaching probabilities of the 5 winning combinations}
\usage{
TopUti(df, rank_pos_1, rank_pos_2, rank_pos_3, rank_pos_4, rank_pos_5)
}
\value{
the reaching probabilities of the 5 winning combinations
}
\description{
Technically this returns the 5 corresponding reaching probabilities of 5 winning combinations.
The maxtrix of exponentiated mean-centred utilities for possible combinations is inputted.
}
\examples{
TopUti()
}
\seealso{

}
